// frontend/src/app/api/model/supplier-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SupplierDto { 
    id?: number;
    name?: string;
    contactPerson?: string;
    email?: string;
    phone?: string;
}